tcp working on simple-rtsp-server

+++ rtsp/src/main/java/com/pedro/rtsp/rtcp/SenderReportTcp.kt
@@ -28,7 +28,7 @@ internal class SenderReportTcp : BaseSenderReport() {
   @Throws(IOException::class)
   private fun sendReportTCP(buffer: ByteArray, channelIdentifier: Byte, type: String, packet: Int, octet: Int, isEnableLogs: Boolean) {
     synchronized(RtpConstants.lock) {
-      tcpHeader[1] = (channelIdentifier + 1).toByte()
+      tcpHeader[1] = (channelIdentifier.toInt() + 1).toByte()
       outputStream?.write(tcpHeader)
       outputStream?.write(buffer, 0, PACKET_LENGTH)
       outputStream?.flush()
